ABM divisions as force directed network

After crawling organizational data, this is used for displaying an interactive graph with a force directed network of the units.

Usage

abm_graph_divisions(
  base_url = "dash/",
  use_size = FALSE,
  link_encoder = NULL,
  links_excluded = NULL,
  link_404 = "",
  prune_graph = FALSE
)

Arguments

base_url

pattern to prefigate links from nodes with in JS click action

use_size

boolean to indicate whether to size bubbles by publication volume, default is FALSE

link_encoder

function to use for encoding outgoing link identifiers, default is NULL but can be set to for example function(x) URLencode(x, reserved = TRUE) to transform outgoing links

links_excluded

a set of node ids for which links will be excluded, default NULL

link_404

relative url to use for excluded links

prune_graph

boolean to indicate if certain non-research nodes should be removed, default: FALSE

Value

a force directed network object from NetworkD3

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 abm_graph_divisions()
 }
} # }