Sync the MS SQL Server database BIBMON to a local SQLite3 db

This function syncs db tables from an mssql source db and writes the data into a local SQLite3 db using buffering, with chunk size set to 1e4 items per chunk, in order to avoid out of memory exceptions when moving large tables.

Usage

db_sync(
  tables_included,
  tables_excluded = c("OA_status", "Document", "LastFailedJobs", "DIVA_School_Dept",
    "Diva_departments", "Doc_statistics", "masterfile_full", "masterfile_2021jan",
    "masterfile_2019dec", "masterfile_2019nov", "masterfile_tmp"),
  overwrite_existing = FALSE
)

Arguments

tables_included

a vector of table names in the source db to be included, by default all tables are included except those excluded

tables_excluded

a vector of table names in the source db to be excluded, by default a number of tables are excluded, specify NULL to not explicitly exclude any tables

overwrite_existing

a logical to indicate whether destination tables should be overwritten if they already exist

Value

invisible result with vector of boolean status flags for synced tables