Calculate average citation indicators across years

This function calculates average indicator values (jcf, cf etc) across a set of years per department, using fractional counting. Data is based on masterfile, using a specific analysis_id (i.e. data version nr.) This three-year average of jcf is used as an performance indicator at KTH.

Usage

mean_indicator_units(
  con,
  starty,
  stopy,
  analysis_level = 2,
  analysis_version_id
)

Arguments

con

connection to db

starty

first publication year of analysis

stopy

last publication year of analysis

analysis_level

organization analysis level. Default is 2 (department).

analysis_version_id

the analysis_id id to be used from masterfile

Value

tibble average citation indicators along with dept name, school name, along with full and fractional publ. counts.